+++ src/package.json
@@ -46,6 +46,7 @@
     "multer": "^1.4.5-lts.1",
     "node-html-parser": "^6.1.11",
     "pdf-parse": "^1.1.1",
+    "pdf2pic": "^3.1.1",
     "qrcode": "^1.5.3",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.2.0"
