+++ src/src/exam/exam.module.ts
@@ -7,11 +7,12 @@ import { Method } from 'src/common/method/method';
 import { CohortMiddleware } from 'src/common/middlewares/cohort.middleware';
 import { UploadFileService } from 'src/upload-file/upload-file.service';
 import { S3Module } from 'src/services/s3/s3.module';
+import { ExamResultPattern } from './exam.result.pattern';
 
 @Module({
 	imports: [HasuraModule, HasuraModuleFromServices, S3Module],
 	controllers: [ExamController],
-	providers: [ExamService, Method, UploadFileService],
+	providers: [ExamService, Method, UploadFileService, ExamResultPattern],
 })
 export class ExamModule implements NestModule {
 	configure(consumer: MiddlewareConsumer) {
